+++ koin/src/jvmMain/kotlin/SingleWithBindsAndRandomQualifier.kt
@@ -1,0 +1,15 @@
+package dev.inmo.micro_utils.koin
+
+import com.benasher44.uuid.uuid4
+import org.koin.core.definition.Definition
+import org.koin.core.instance.InstanceFactory
+import org.koin.core.module.Module
+import kotlin.reflect.KClass
+
+inline fun <reified T : Any> Module.singleWithRandomQualifierAndBinds(
+    createdAtStart: Boolean = false,
+    bindFilter: (KClass<*>) -> Boolean = { true },
+    noinline definition: Definition<T>
+): Pair<Module, InstanceFactory<*>> {
+    return singleWithBinds(uuid4().toString(), createdAtStart, bindFilter, definition)
+}